move SessionGuard to src/types
